show[Semantics] now take a semantic context
It will now typecheck the tree first, and only then print out the attributes.
There's newly introduced show[Attributes] that doesn't do the typechecking part.
